Move snippet sources
