+++ jms/src/test/java/docs/javadsl/JmsBufferedAckConnectorsTest.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.jms.javadsl;
+package docs.javadsl;
 
 import akka.Done;
 import akka.actor.ActorSystem;
@@ -10,6 +10,9 @@ import akka.japi.Pair;
 import akka.stream.ActorMaterializer;
 import akka.stream.Materializer;
 import akka.stream.alpakka.jms.*;
+import akka.stream.alpakka.jms.javadsl.JmsConsumer;
+import akka.stream.alpakka.jms.javadsl.JmsConsumerControl;
+import akka.stream.alpakka.jms.javadsl.JmsProducer;
 import akka.stream.javadsl.Sink;
 import akka.stream.javadsl.Source;
 import akka.testkit.javadsl.TestKit;
@@ -33,9 +36,8 @@ import java.util.stream.Stream;
 
 import static org.junit.Assert.assertEquals;
 
-public class JmsTxConnectorsTest {
+public class JmsBufferedAckConnectorsTest {
 
-  // #create-test-message-list
   private List<JmsTextMessage> createTestMessageList() {
     List<Integer> intsIn = Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
     List<JmsTextMessage> msgsIn = new ArrayList<>();
@@ -50,48 +52,36 @@ public class JmsTxConnectorsTest {
 
     return msgsIn;
   }
-  // #create-test-message-list
 
   @Test
   public void publishAndConsume() throws Exception {
     withServer(
         ctx -> {
-          // #connection-factory
-          //            ActiveMQConnectionFactory connectionFactory = new
-          // ActiveMQConnectionFactory(ctx.url);
           ConnectionFactory connectionFactory = new ActiveMQConnectionFactory(ctx.url);
-          // #connection-factory
 
-          // #create-text-sink
           Sink<String, CompletionStage<Done>> jmsSink =
               JmsProducer.textSink(JmsProducerSettings.create(connectionFactory).withQueue("test"));
-          // #create-text-sink
 
-          // #run-text-sink
           List<String> in = Arrays.asList("a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k");
           Source.from(in).runWith(jmsSink, materializer);
-          // #run-text-sink
 
-          // #create-text-source
-          Source<TxEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.txSource(
+          Source<AckEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.ackSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
+                      .withBufferSize(5)
                       .withQueue("test"));
-          // #create-text-source
 
-          // #run-text-source
           CompletionStage<List<String>> result =
               jmsSource
                   .take(in.size())
                   .map(env -> new Pair<>(env, ((TextMessage) env.message()).getText()))
                   .map(
                       pair -> {
-                        pair.first().commit();
+                        pair.first().acknowledge();
                         return pair.second();
                       })
                   .runWith(Sink.seq(), materializer);
-          // #run-text-source
           List<String> out = new ArrayList<>(result.toCompletableFuture().get(3, TimeUnit.SECONDS));
           Collections.sort(out);
           assertEquals(in, out);
@@ -104,25 +94,19 @@ public class JmsTxConnectorsTest {
         ctx -> {
           ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory(ctx.url);
 
-          // #create-jms-sink
           Sink<JmsTextMessage, CompletionStage<Done>> jmsSink =
               JmsProducer.create(JmsProducerSettings.create(connectionFactory).withQueue("test"));
-          // #create-jms-sink
 
-          // #create-messages-with-properties
           List<JmsTextMessage> msgsIn = createTestMessageList();
-          // #create-messages-with-properties
 
-          // #run-jms-sink
           Source.from(msgsIn).runWith(jmsSink, materializer);
-          // #run-jms-sink
 
           // #create-jms-source
-          Source<TxEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.txSource(
+          Source<AckEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.ackSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
-                      .withAckTimeout(1, TimeUnit.SECONDS)
+                      .withBufferSize(5)
                       .withQueue("test"));
           // #create-jms-source
 
@@ -132,7 +116,7 @@ public class JmsTxConnectorsTest {
                   .take(msgsIn.size())
                   .map(
                       env -> {
-                        env.commit();
+                        env.acknowledge();
                         return env.message();
                       })
                   .runWith(Sink.seq(), materializer);
@@ -171,12 +155,9 @@ public class JmsTxConnectorsTest {
         ctx -> {
           ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory(ctx.url);
 
-          // #create-jms-sink
           Sink<JmsTextMessage, CompletionStage<Done>> jmsSink =
               JmsProducer.create(JmsProducerSettings.create(connectionFactory).withQueue("test"));
-          // #create-jms-sink
 
-          // #create-messages-with-properties
           List<JmsTextMessage> msgsIn =
               createTestMessageList()
                   .stream()
@@ -186,31 +167,25 @@ public class JmsTxConnectorsTest {
                           jmsTextMessage.withHeader(JmsCorrelationId.create("correlationId")))
                   .map(jmsTextMessage -> jmsTextMessage.withHeader(JmsReplyTo.queue("test-reply")))
                   .collect(Collectors.toList());
-          // #create-messages-with-properties
 
-          // #run-jms-sink
           Source.from(msgsIn).runWith(jmsSink, materializer);
-          // #run-jms-sink
 
-          // #create-jms-source
-          Source<TxEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.txSource(
+          Source<AckEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.ackSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
+                      .withBufferSize(5)
                       .withQueue("test"));
-          // #create-jms-source
 
-          // #run-jms-source
           CompletionStage<List<Message>> result =
               jmsSource
                   .take(msgsIn.size())
                   .map(
                       env -> {
-                        env.commit();
+                        env.acknowledge();
                         return env.message();
                       })
                   .runWith(Sink.seq(), materializer);
-          // #run-jms-source
 
           List<Message> outMessages =
               new ArrayList<>(result.toCompletableFuture().get(3, TimeUnit.SECONDS));
@@ -254,16 +229,14 @@ public class JmsTxConnectorsTest {
 
           Source.from(msgsIn).runWith(jmsSink, materializer);
 
-          // #create-jms-source-with-selector
-          Source<TxEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.txSource(
+          Source<AckEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.ackSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
+                      .withBufferSize(5)
                       .withQueue("test")
                       .withSelector("IsOdd = TRUE"));
-          // #create-jms-source-with-selector
 
-          // #assert-only-odd-messages-received
           List<JmsTextMessage> oddMsgsIn =
               msgsIn
                   .stream()
@@ -276,7 +249,7 @@ public class JmsTxConnectorsTest {
                   .take(oddMsgsIn.size())
                   .map(
                       env -> {
-                        env.commit();
+                        env.acknowledge();
                         return env.message();
                       })
                   .runWith(Sink.seq(), materializer);
@@ -306,7 +279,6 @@ public class JmsTxConnectorsTest {
             assertEquals(1, outMsg.getIntProperty("Number") % 2);
             msgIdx++;
           }
-          // #assert-only-odd-messages-received
         });
   }
 
@@ -320,46 +292,42 @@ public class JmsTxConnectorsTest {
           List<String> inNumbers =
               IntStream.range(0, 10).boxed().map(String::valueOf).collect(Collectors.toList());
 
-          // #create-topic-sink
           Sink<String, CompletionStage<Done>> jmsTopicSink =
               JmsProducer.textSink(
                   JmsProducerSettings.create(connectionFactory).withTopic("topic"));
-          // #create-topic-sink
           Sink<String, CompletionStage<Done>> jmsTopicSink2 =
               JmsProducer.textSink(
                   JmsProducerSettings.create(connectionFactory).withTopic("topic"));
 
-          // #create-topic-source
-          Source<TxEnvelope, JmsConsumerControl> jmsTopicSource =
-              JmsConsumer.txSource(
+          Source<AckEnvelope, JmsConsumerControl> jmsTopicSource =
+              JmsConsumer.ackSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(1)
+                      .withBufferSize(5)
                       .withTopic("topic"));
-          // #create-topic-source
-          Source<TxEnvelope, JmsConsumerControl> jmsTopicSource2 =
-              JmsConsumer.txSource(
+          Source<AckEnvelope, JmsConsumerControl> jmsTopicSource2 =
+              JmsConsumer.ackSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(1)
+                      .withBufferSize(5)
                       .withTopic("topic"));
 
-          // #run-topic-source
           CompletionStage<List<String>> result =
               jmsTopicSource
                   .take(in.size() + inNumbers.size())
                   .map(
                       env -> {
-                        env.commit();
+                        env.acknowledge();
                         return ((TextMessage) env.message()).getText();
                       })
                   .runWith(Sink.seq(), materializer)
                   .thenApply(l -> l.stream().sorted().collect(Collectors.toList()));
-          // #run-topic-source
           CompletionStage<List<String>> result2 =
               jmsTopicSource2
                   .take(in.size() + inNumbers.size())
                   .map(
                       env -> {
-                        env.commit();
+                        env.acknowledge();
                         return ((TextMessage) env.message()).getText();
                       })
                   .runWith(Sink.seq(), materializer)
@@ -367,9 +335,7 @@ public class JmsTxConnectorsTest {
 
           Thread.sleep(500);
 
-          // #run-topic-sink
           Source.from(in).runWith(jmsTopicSink, materializer);
-          // #run-topic-sink
           Source.from(inNumbers).runWith(jmsTopicSink2, materializer);
 
           assertEquals(
