+++ jms/src/test/java/docs/javadsl/JmsTxConnectorsTest.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.jms.javadsl;
+package docs.javadsl;
 
 import akka.Done;
 import akka.actor.ActorSystem;
@@ -10,6 +10,9 @@ import akka.japi.Pair;
 import akka.stream.ActorMaterializer;
 import akka.stream.Materializer;
 import akka.stream.alpakka.jms.*;
+import akka.stream.alpakka.jms.javadsl.JmsConsumer;
+import akka.stream.alpakka.jms.javadsl.JmsConsumerControl;
+import akka.stream.alpakka.jms.javadsl.JmsProducer;
 import akka.stream.javadsl.Sink;
 import akka.stream.javadsl.Source;
 import akka.testkit.javadsl.TestKit;
@@ -33,8 +36,9 @@ import java.util.stream.Stream;
 
 import static org.junit.Assert.assertEquals;
 
-public class JmsBufferedAckConnectorsTest {
+public class JmsTxConnectorsTest {
 
+  // #create-test-message-list
   private List<JmsTextMessage> createTestMessageList() {
     List<Integer> intsIn = Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
     List<JmsTextMessage> msgsIn = new ArrayList<>();
@@ -49,36 +53,48 @@ public class JmsBufferedAckConnectorsTest {
 
     return msgsIn;
   }
+  // #create-test-message-list
 
   @Test
   public void publishAndConsume() throws Exception {
     withServer(
         ctx -> {
+          // #connection-factory
+          //            ActiveMQConnectionFactory connectionFactory = new
+          // ActiveMQConnectionFactory(ctx.url);
           ConnectionFactory connectionFactory = new ActiveMQConnectionFactory(ctx.url);
+          // #connection-factory
 
+          // #create-text-sink
           Sink<String, CompletionStage<Done>> jmsSink =
               JmsProducer.textSink(JmsProducerSettings.create(connectionFactory).withQueue("test"));
+          // #create-text-sink
 
+          // #run-text-sink
           List<String> in = Arrays.asList("a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k");
           Source.from(in).runWith(jmsSink, materializer);
+          // #run-text-sink
 
-          Source<AckEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.ackSource(
+          // #create-text-source
+          Source<TxEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.txSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
-                      .withBufferSize(5)
                       .withQueue("test"));
+          // #create-text-source
 
+          // #run-text-source
           CompletionStage<List<String>> result =
               jmsSource
                   .take(in.size())
                   .map(env -> new Pair<>(env, ((TextMessage) env.message()).getText()))
                   .map(
                       pair -> {
-                        pair.first().acknowledge();
+                        pair.first().commit();
                         return pair.second();
                       })
                   .runWith(Sink.seq(), materializer);
+          // #run-text-source
           List<String> out = new ArrayList<>(result.toCompletableFuture().get(3, TimeUnit.SECONDS));
           Collections.sort(out);
           assertEquals(in, out);
@@ -91,19 +107,25 @@ public class JmsBufferedAckConnectorsTest {
         ctx -> {
           ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory(ctx.url);
 
+          // #create-jms-sink
           Sink<JmsTextMessage, CompletionStage<Done>> jmsSink =
               JmsProducer.create(JmsProducerSettings.create(connectionFactory).withQueue("test"));
+          // #create-jms-sink
 
+          // #create-messages-with-properties
           List<JmsTextMessage> msgsIn = createTestMessageList();
+          // #create-messages-with-properties
 
+          // #run-jms-sink
           Source.from(msgsIn).runWith(jmsSink, materializer);
+          // #run-jms-sink
 
           // #create-jms-source
-          Source<AckEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.ackSource(
+          Source<TxEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.txSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
-                      .withBufferSize(5)
+                      .withAckTimeout(1, TimeUnit.SECONDS)
                       .withQueue("test"));
           // #create-jms-source
 
@@ -113,7 +135,7 @@ public class JmsBufferedAckConnectorsTest {
                   .take(msgsIn.size())
                   .map(
                       env -> {
-                        env.acknowledge();
+                        env.commit();
                         return env.message();
                       })
                   .runWith(Sink.seq(), materializer);
@@ -152,9 +174,12 @@ public class JmsBufferedAckConnectorsTest {
         ctx -> {
           ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory(ctx.url);
 
+          // #create-jms-sink
           Sink<JmsTextMessage, CompletionStage<Done>> jmsSink =
               JmsProducer.create(JmsProducerSettings.create(connectionFactory).withQueue("test"));
+          // #create-jms-sink
 
+          // #create-messages-with-properties
           List<JmsTextMessage> msgsIn =
               createTestMessageList()
                   .stream()
@@ -164,25 +189,31 @@ public class JmsBufferedAckConnectorsTest {
                           jmsTextMessage.withHeader(JmsCorrelationId.create("correlationId")))
                   .map(jmsTextMessage -> jmsTextMessage.withHeader(JmsReplyTo.queue("test-reply")))
                   .collect(Collectors.toList());
+          // #create-messages-with-properties
 
+          // #run-jms-sink
           Source.from(msgsIn).runWith(jmsSink, materializer);
+          // #run-jms-sink
 
-          Source<AckEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.ackSource(
+          // #create-jms-source
+          Source<TxEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.txSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
-                      .withBufferSize(5)
                       .withQueue("test"));
+          // #create-jms-source
 
+          // #run-jms-source
           CompletionStage<List<Message>> result =
               jmsSource
                   .take(msgsIn.size())
                   .map(
                       env -> {
-                        env.acknowledge();
+                        env.commit();
                         return env.message();
                       })
                   .runWith(Sink.seq(), materializer);
+          // #run-jms-source
 
           List<Message> outMessages =
               new ArrayList<>(result.toCompletableFuture().get(3, TimeUnit.SECONDS));
@@ -226,14 +257,16 @@ public class JmsBufferedAckConnectorsTest {
 
           Source.from(msgsIn).runWith(jmsSink, materializer);
 
-          Source<AckEnvelope, JmsConsumerControl> jmsSource =
-              JmsConsumer.ackSource(
+          // #create-jms-source-with-selector
+          Source<TxEnvelope, JmsConsumerControl> jmsSource =
+              JmsConsumer.txSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(5)
-                      .withBufferSize(5)
                       .withQueue("test")
                       .withSelector("IsOdd = TRUE"));
+          // #create-jms-source-with-selector
 
+          // #assert-only-odd-messages-received
           List<JmsTextMessage> oddMsgsIn =
               msgsIn
                   .stream()
@@ -246,7 +279,7 @@ public class JmsBufferedAckConnectorsTest {
                   .take(oddMsgsIn.size())
                   .map(
                       env -> {
-                        env.acknowledge();
+                        env.commit();
                         return env.message();
                       })
                   .runWith(Sink.seq(), materializer);
@@ -276,6 +309,7 @@ public class JmsBufferedAckConnectorsTest {
             assertEquals(1, outMsg.getIntProperty("Number") % 2);
             msgIdx++;
           }
+          // #assert-only-odd-messages-received
         });
   }
 
@@ -289,42 +323,46 @@ public class JmsBufferedAckConnectorsTest {
           List<String> inNumbers =
               IntStream.range(0, 10).boxed().map(String::valueOf).collect(Collectors.toList());
 
+          // #create-topic-sink
           Sink<String, CompletionStage<Done>> jmsTopicSink =
               JmsProducer.textSink(
                   JmsProducerSettings.create(connectionFactory).withTopic("topic"));
+          // #create-topic-sink
           Sink<String, CompletionStage<Done>> jmsTopicSink2 =
               JmsProducer.textSink(
                   JmsProducerSettings.create(connectionFactory).withTopic("topic"));
 
-          Source<AckEnvelope, JmsConsumerControl> jmsTopicSource =
-              JmsConsumer.ackSource(
+          // #create-topic-source
+          Source<TxEnvelope, JmsConsumerControl> jmsTopicSource =
+              JmsConsumer.txSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(1)
-                      .withBufferSize(5)
                       .withTopic("topic"));
-          Source<AckEnvelope, JmsConsumerControl> jmsTopicSource2 =
-              JmsConsumer.ackSource(
+          // #create-topic-source
+          Source<TxEnvelope, JmsConsumerControl> jmsTopicSource2 =
+              JmsConsumer.txSource(
                   JmsConsumerSettings.create(connectionFactory)
                       .withSessionCount(1)
-                      .withBufferSize(5)
                       .withTopic("topic"));
 
+          // #run-topic-source
           CompletionStage<List<String>> result =
               jmsTopicSource
                   .take(in.size() + inNumbers.size())
                   .map(
                       env -> {
-                        env.acknowledge();
+                        env.commit();
                         return ((TextMessage) env.message()).getText();
                       })
                   .runWith(Sink.seq(), materializer)
                   .thenApply(l -> l.stream().sorted().collect(Collectors.toList()));
+          // #run-topic-source
           CompletionStage<List<String>> result2 =
               jmsTopicSource2
                   .take(in.size() + inNumbers.size())
                   .map(
                       env -> {
-                        env.acknowledge();
+                        env.commit();
                         return ((TextMessage) env.message()).getText();
                       })
                   .runWith(Sink.seq(), materializer)
@@ -332,7 +370,9 @@ public class JmsBufferedAckConnectorsTest {
 
           Thread.sleep(500);
 
+          // #run-topic-sink
           Source.from(in).runWith(jmsTopicSink, materializer);
+          // #run-topic-sink
           Source.from(inNumbers).runWith(jmsTopicSink2, materializer);
 
           assertEquals(
